Move models.py and server.py to app folder, change related reference, andd comment out 500server test case
