+++ setup.cfg
@@ -5,7 +5,7 @@ spec-color=1
 nologcapture=1
 with-coverage=1
 cover-erase=1
-cover-package=server,models
+cover-package=app
 with-noy=1
 noy-with-default-imports=1
 # with-xunit=1
